+++ apps/backend/prisma/prisma.ts
@@ -1,8 +1,8 @@
 import { PrismaClient } from '@prisma/client/edge';
 import { withAccelerate } from '@prisma/extension-accelerate';
 //👷開発用
-// import { config } from 'dotenv';
-// config();
+import { config } from 'dotenv';
+config();
 
 const DATABASE_URL = process.env.DATABASE_URL;
