+++ apps/backend/prisma/schema.prisma
@@ -9,34 +9,27 @@ datasource db {
 }
 
 model User {
-  id            Int      @id @default(autoincrement())
-  clerk_id      String   @unique
-  line_id       String?  @unique
-  user_name     String?  @unique
-  display_name  String
-  profile_image String
-  target        String
-  is_active     Boolean  @default(true)
-  created_at    DateTime @default(now())
-  updated_at    DateTime @updatedAt
-
-  // リレーション
+  id                   Int          @id @default(autoincrement())
+  display_name         String
+  profile_image        String
+  created_at           DateTime     @default(now())
+  updated_at           DateTime     @updatedAt
+  clerk_id             String       @unique
+  user_name            String?      @unique
+  line_id              String?      @unique
   activities           Activity[]
-  weekly_targets       WeeklyTarget[]
-  sent_friendships     Friendship[]   @relation("SentFriendships")
-  received_friendships Friendship[]   @relation("ReceivedFriendships")
+  received_friendships Friendship[] @relation("ReceivedFriendships")
+  sent_friendships     Friendship[] @relation("SentFriendships")
 
   @@map("users")
 }
 
 model Activity {
-  id            Int      @id @default(autoincrement())
-  user_clerk_id String
-  activity_date DateTime
-  created_at    DateTime @default(now())
-  updated_at    DateTime @updatedAt
-
-  // リレーション
+  id               Int              @id @default(autoincrement())
+  activity_date    DateTime
+  created_at       DateTime         @default(now())
+  updated_at       DateTime         @updatedAt
+  user_clerk_id    String
   user             User             @relation(fields: [user_clerk_id], references: [clerk_id], onDelete: Cascade)
   activity_details ActivityDetail[]
 
@@ -52,34 +45,26 @@ model ActivityDetail {
   category         String
   created_at       DateTime @default(now())
   updated_at       DateTime @updatedAt
-
-  // リレーション
-  activity Activity @relation(fields: [activity_id], references: [id], onDelete: Cascade)
+  activity         Activity @relation(fields: [activity_id], references: [id], onDelete: Cascade)
 
   @@index([activity_id])
   @@map("activity_details")
 }
 
-enum WeeklyTargetStatus {
-  pending
-  started
-  completed
-}
+model Friendship {
+  id                Int              @id @default(autoincrement())
+  status            FriendshipStatus @default(pending)
+  created_at        DateTime         @default(now())
+  updated_at        DateTime         @updatedAt
+  receiver_clerk_id String
+  sender_clerk_id   String
+  receiver          User             @relation("ReceivedFriendships", fields: [receiver_clerk_id], references: [clerk_id], onDelete: Cascade)
+  sender            User             @relation("SentFriendships", fields: [sender_clerk_id], references: [clerk_id], onDelete: Cascade)
 
-model WeeklyTarget {
-  id                Int                @id @default(autoincrement())
-  user_clerk_id     String
-  title             String
-  description       String
-  status            WeeklyTargetStatus @default(pending)
-  created_at        DateTime           @default(now())
-  updated_at        DateTime           @updatedAt
-
-  // リレーション
-  user User @relation(fields: [user_clerk_id], references: [clerk_id], onDelete: Cascade)
-
-  @@index([user_clerk_id])
-  @@map("weekly_targets")
+  @@unique([sender_clerk_id, receiver_clerk_id])
+  @@index([sender_clerk_id])
+  @@index([receiver_clerk_id])
+  @@map("friendships")
 }
 
 enum FriendshipStatus {
@@ -88,20 +73,8 @@ enum FriendshipStatus {
   rejected
 }
 
-model Friendship {
-  id                Int              @id @default(autoincrement())
-  sender_clerk_id   String
-  receiver_clerk_id String
-  status            FriendshipStatus @default(pending)
-  created_at        DateTime         @default(now())
-  updated_at        DateTime         @updatedAt
-
-  // リレーション
-  sender   User @relation("SentFriendships", fields: [sender_clerk_id], references: [clerk_id], onDelete: Cascade)
-  receiver User @relation("ReceivedFriendships", fields: [receiver_clerk_id], references: [clerk_id], onDelete: Cascade)
-
-  @@unique([sender_clerk_id, receiver_clerk_id])
-  @@index([sender_clerk_id])
-  @@index([receiver_clerk_id])
-  @@map("friendships")
+enum WeeklyTargetStatus {
+  pending
+  started
+  completed
 }
