+++ apps/frontend/src/_components/manual_ui/activity_parts/activityButton/ActiveButton.tsx
@@ -6,7 +6,7 @@ import { sendToFriends } from './action';
 export default function ActiveButton({ userId }: { userId: string }) {
   return (
     <div className="flex flex-col items-center justify-center bg-slate-800/30 border border-slate-700/50 rounded-xl p-5 shadow-xl hover:shadow-blue-900/10 transition-all duration-300">
-      <p>活動時間はカレンダーの日付から登録</p>
+      <p>詳細な時間は日付から個別で入力</p>
       <div className="flex flex-col justify-center items-center">
         <Button
           onClick={() => {
