+++ apps/frontend/src/app/(Pages)/friends/[clerk_id]/page.tsx
@@ -1,21 +1,12 @@
 import { ActivityTabs } from '@/_components/manual_ui/activity_parts/ActivityTabs';
 import { PageHeader } from '@/_components/manual_ui/activity_parts/PageHeader';
-import { client } from '@/utils/client';
 import { User } from 'types/type';
+import { getFriendsData } from '../function';
 
 export default async function FriendsInfo({ params }: { params: Promise<{ clerk_id: string }> }) {
   const { clerk_id } = await params;
 
-  const res = await client.api.user.friends[':clerk_id'].$get({
-    param: { clerk_id: clerk_id },
-  });
-
-  const friendsData = (await res.json()) as User;
-  console.log(
-    'フレンド詳細ページ - 取得したユーザー:',
-    friendsData.display_name,
-    friendsData.clerk_id
-  );
+  const friendsData = await getFriendsData(clerk_id) as User;
 
   return (
     <div className="min-h-screen text-white">
