+++ apps/frontend/src/app/(Pages)/friends/deleteButton.tsx
@@ -1,10 +1,7 @@
-'use client';
-
+"use client"
 import { Button } from '@/_components/shadcn_ui/button';
-import { client } from '@/utils/client';
-import { useRouter } from 'next/navigation';
+import { deleteFriend } from './function';
 import { useState } from 'react';
-import { toast } from 'sonner';
 
 const DeleteButton = ({
   friendClerkId,
@@ -13,45 +10,23 @@ const DeleteButton = ({
   friendClerkId: string;
   myClerkId: string;
 }) => {
-  const router = useRouter();
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleDelete = async () => {
-    if (isLoading) return;
-
-    try {
-      setIsLoading(true);
-
-      // 直接APIを呼び出す
-      const response = await client.api.friendRequest.deleteFriend.$delete({
-        json: { myClerkId, friendClerkId },
-      });
-
-      if (!response.ok) {
-        throw new Error('友達の削除に失敗しました');
-      }
-      toast.success('友達を削除しました');
-
-      // ページを更新する
-      router.refresh();
-    } catch (error) {
-      console.error('削除エラー:', error);
-      toast.error('削除に失敗しました');
-    } finally {
-      setIsLoading(false);
-    }
-  };
+const [isDeleting, setIsDeleting] = useState(false);
 
   return (
     <div className="ml-2">
       <Button
-        onClick={handleDelete}
+        onClick={async () => {
+          setIsDeleting(true);
+          await deleteFriend(myClerkId, friendClerkId);
+          setIsDeleting(false);
+        }}
         type="button"
         variant="destructive"
         size="sm"
-        disabled={isLoading}
+        disabled={isDeleting}
+        className="cursor-pointer"
       >
-        {isLoading ? '...' : '削除'}
+        {isDeleting ? '削除中...' : '削除'}
       </Button>
     </div>
   );
